+++ src/main/java/com/qdw/calendaing/schedulerStregys/lp/constraintGenerater/DefaultConstraintGenerater.java
@@ -19,13 +19,13 @@ public class DefaultConstraintGenerater extends AbstractConstraintGenerater {
 
 
     @Override
-    public List<List<Integer>> generateOne(NetContext netContext, Collection<Flow> flows, ConstraintType constraintType) {
+    public List<List<Double>> generateOne(NetContext netContext, Collection<Flow> flows, ConstraintType constraintType) {
         if (netContext.getNetwork()==null || netContext.getRequirements()==null || constraintType==null){
             log.error("无效参数");
             return new ArrayList<>();
         }
 
-        List<List<Integer>> list = new ArrayList<>();
+        List<List<Double>> list = new ArrayList<>();
         switch (constraintType){
             case RONGLIANG:
                 list = getRLCons(netContext,flows);
@@ -44,29 +44,29 @@ public class DefaultConstraintGenerater extends AbstractConstraintGenerater {
 
     // 1< 2> 3=
     // 获取链路容量约束
-    private List<List<Integer>> getRLCons(NetContext netContext,Collection<Flow> flows){
+    private List<List<Double>> getRLCons(NetContext netContext,Collection<Flow> flows){
         Requirements requirements = netContext.getRequirements();
         Network network = netContext.getNetwork();
-        List<List<Integer>> res = new ArrayList<>(flows.size());
+        List<List<Double>> res = new ArrayList<>(flows.size());
         System.out.println("linksdesize:"+network.getLinks().size());
         for (int i = requirements.getEarliestSlot(); i <= requirements.getLatestSlot(); i++) {
             for (Link link : network.getLinks().values()) {
-                List<Integer> list = new LinkedList<>();
+                List<Double> list = new LinkedList<>();
                 boolean flag = false;
                 for (Flow flow : flows) {
                     if (flow.getTimeSlot()==i && flow.isCover(link)) {
 //                        System.out.println("###link:"+link.getId());
-                        list.add(1);
+                        list.add(1.0);
                         flag = true;
                     } else {
-                        list.add(0);
+                        list.add(0.0);
                     }
                 }
                 if (!flag){
                     continue;
                 }
-                list.add(1);
-                list.add((int) link.getLinkInfo(i).getResidualCapacity());
+                list.add(1.0);
+                list.add(link.getLinkInfo(i).getResidualCapacity());
                 res.add(list);
             }
         }
@@ -77,26 +77,26 @@ public class DefaultConstraintGenerater extends AbstractConstraintGenerater {
 
     }
     // 获取流量约束
-    private List<List<Integer>> getLLCons(NetContext netContext,Collection<Flow> flows){
+    private List<List<Double>> getLLCons(NetContext netContext,Collection<Flow> flows){
         int flowsOfAll = flows.size();
         System.out.println("所有初始流的数量为:"+flowsOfAll);
 
-        List<List<Integer>> res = new LinkedList<>();
+        List<List<Double>> res = new LinkedList<>();
         int j = 0;
         for (int i = 0; i < flowsOfAll; i++) {
 
-            List<Integer> list = new ArrayList<>(flowsOfAll+2);
+            List<Double> list = new ArrayList<>(flowsOfAll+2);
             for (int k = 0; k < flowsOfAll; k++) {
                 if (k==j){
-                    list.add(1);
+                    list.add(1.0);
                 }else {
-                    list.add(0);
+                    list.add(0.0);
                 }
             }
             j++;
 //            list.set(i,2);
-            list.add(2);
-            list.add(0);
+            list.add(2.0);
+            list.add(0.0);
             res.add(list);
         }
         System.out.println("参数个数为：" + res.get(0).size());
@@ -105,34 +105,34 @@ public class DefaultConstraintGenerater extends AbstractConstraintGenerater {
     }
 
     // 获取请求的需求约束
-    private List<List<Integer>> getXQCons(NetContext netContext,Collection<Flow> flows){
+    private List<List<Double>> getXQCons(NetContext netContext,Collection<Flow> flows){
         Requirements requirements = netContext.getRequirements();
-        List<List<Integer>> res = new LinkedList<>();
+        List<List<Double>> res = new LinkedList<>();
         int flowsOfAll = requirements.getFlowsOfAll();
         System.out.println("所有初始流的数量为:"+flowsOfAll);
         int i = 0;
-        Map<Requirements.Requirement,List<Integer>> map = new LinkedHashMap<>();
+        Map<Requirements.Requirement,List<Double>> map = new LinkedHashMap<>();
         for (Flow flow : flows) {
             if (!map.containsKey(flow.getThisR())){
-                List<Integer> list = new LinkedList<>();
+                List<Double> list = new LinkedList<>();
                 for (int j = 0; j < i; j++) {
-                    list.add(0);
+                    list.add(0.0);
                 }
-                list.add(1);
+                list.add(1.0);
                 map.put(flow.getThisR(),list);
             }else {
-                List<Integer> list = map.get(flow.getThisR());
-                list.add(1);
+                List<Double> list = map.get(flow.getThisR());
+                list.add(1.0);
             }
             i++;
         }
         for (Requirements.Requirement requirement : map.keySet()) {
-            List<Integer> list = map.get(requirement);
+            List<Double> list = map.get(requirement);
             while (list.size()<flowsOfAll){
-                list.add(0);
+                list.add(0.0);
             }
-            list.add(3);
-            list.add((int)(requirement.getDemand()-requirement.getMeetDemand()));
+            list.add(3.0);
+            list.add((requirement.getDemand()-requirement.getMeetDemand()));
             res.add(list);
         }
         System.out.println("参数个数为：" + res.get(0).size());
@@ -144,10 +144,10 @@ public class DefaultConstraintGenerater extends AbstractConstraintGenerater {
 
     // 总流量最大，除了虚拟流
     @Override
-    public List<Integer> getObjFunc(NetContext netContext,Collection<Flow> flows) {
-        List<Integer> res = new LinkedList<>();
+    public List<Double> getObjFunc(NetContext netContext,Collection<Flow> flows) {
+        List<Double> res = new LinkedList<>();
         for (Flow flow : flows) {
-            res.add(getCost(flow));
+            res.add((double)getCost(flow));
         }
 
         return res;
